requirement.txt
keyboard
pynput-1.7.6

